# Disable tracking for perf test