This is an analysis of the impact of usage metadata tracking we added in https://github.com/metabase/metabase/issues/38229
In this branch, I disabled all the tracking and compare it against master(with tracking enabled)

Note: the tracking version only includes tracking of: view_count, report_dashboard.last_viewed_at, report_card.last_used_at. Field usage is already disabled on master.

Test spec: https://www.notion.so/metabase/Perf-Testing-Plan-Usage-data-tracking-5c9a6b7cf1444d97b338437ef673d12b?pvs=4

### Method
- The postgres DB is created with this [sql](https://gist.github.com/qnkhuat/1dc6af93365875ec372ce01936340104)
- Dashboard is created using this script: https://gist.github.com/qnkhuat/23552c5defc9db007de83f463a85e681
it creates a dashboard with 20 cards, each is a simple query against the table from the above DB
- Create 1 admin user (crowberto) and 1 sandboxed user (rasta)
- Spinning up each metabase jar with `java -Xms512m -Xmx8196m -jar {{JAR_FILE}}`
- Monitor using visual VM
- Stress testing using k6 with this [script](https://gist.github.com/qnkhuat/2b0ddf64b2b5c339f20ccf7bb53d02b6). It gets the dashboard, query_metadata and executes all the cards. Setup with a constant 8 users for 10mins. 
- Out of 10 users, 8 will be admins, 2 will be sandboxed users and rotate sign-in.

I did this a few times with each jar and saw consistent results.

Conclusion: I don't see much difference in CPU and memory usage on both master and 50. The # interations from k6 are quite the same too.

## 50: tracking vs no tracking

Viewing the dashboard makes:
- 6040 db calls in total with tracking
- 6012 db calls without tracking

### Postgres
no tracking (3bcd4a4b57a53b8a17ce82746d05f99baf65e495)
<img width="1496" alt="Screenshot 2024-08-06 at 18 14 15 (2)" src="https://github.com/user-attachments/assets/0f121b0e-3c21-485b-9429-079ea09e1374">


with tracking  (branch: [50-with-tracking](https://github.com/metabase/metabase/compare/release-x.50.x...50-with-tracking?expand=1))
<img width="1496" alt="Screenshot 2024-08-06 at 18 02 59 (2)" src="https://github.com/user-attachments/assets/ce184a12-1f9a-4102-bdb1-9cff1265830f">


### Mysql
no tracking (3bcd4a4b57a53b8a17ce82746d05f99baf65e495)

<img width="1496" alt="Screenshot 2024-08-06 at 19 19 10 (2)" src="https://github.com/user-attachments/assets/06d5c0df-2775-4a34-b367-f1541a0214c1">


with tracking (branch: [50-with-tracking](https://github.com/metabase/metabase/compare/release-x.50.x...50-with-tracking?expand=1))
<img width="1496" alt="Screenshot 2024-08-06 at 19 08 05 (2)" src="https://github.com/user-attachments/assets/cf7aea9c-9c07-460b-a811-49c32071c07d">


# API calls
| Scenario          | Endpoint                                          | Command                                                                                                    | Siege Result                                                                                           |
|-------------------|---------------------------------------------------|------------------------------------------------------------------------------------------------------------|--------------------------------------------------------------------------------------------------------|
| **Admin - With Tracking** | `/api/dashboard/10`                                | `siege --concurrent=15 --time=60s --header="Cookie: $ADMIN" "http://localhost:3000/api/dashboard/10"` | Transactions: 85 hits<br>Availability: 100.00 %<br>Elapsed time: 60.06 secs<br>Response time: 10.04 secs<br>Transaction rate: 1.42 trans/sec<br>Throughput: 1.97 MB/sec<br>Longest transaction: 18.76<br>Shortest transaction: 7.08 |
| **Admin - Without Tracking** | `/api/dashboard/10`                                | `siege --concurrent=15 --time=60s --header="Cookie: $ADMIN" "http://localhost:3000/api/dashboard/10"| Transactions: 86 hits<br>Availability: 100.00 %<br>Elapsed time: 60.11 secs<br>Response time: 9.59 secs<br>Transaction rate: 1.43 trans/sec<br>Throughput: 1.99 MB/sec<br>Longest transaction: 12.68<br>Shortest transaction: 7.73 |
| **Admin - With Tracking** | `/api/dashboard/10/query_metadata`                  | `siege --concurrent=15 --time=60s --header="Cookie: $ADMIN" "http://localhost:3000/api/dashboard/10/query_metadata"` | Transactions: 30 hits<br>Availability: 100.00 %<br>Elapsed time: 60.80 secs<br>Response time: 22.34 secs<br>Transaction rate: 0.49 trans/sec<br>Throughput: 0.26 MB/sec<br>Longest transaction: 25.01<br>Shortest transaction: 19.02 |
| **Admin - Without Tracking** | `/api/dashboard/10/query_metadata`                  | `siege --concurrent=15 --time=60s --header="Cookie: $ADMIN" "http://localhost:3000/api/dashboard/10/query_metadata"` | Transactions: 30 hits<br>Availability: 100.00 %<br>Elapsed time: 60.25 secs<br>Response time: 21.98 secs<br>Transaction rate: 0.50 trans/sec<br>Throughput: 0.27 MB/sec<br>Longest transaction: 24.12<br>Shortest transaction: 18.69 |
| **Sandboxed - With Tracking** | `/api/dashboard/10`                                | `siege --concurrent=15 --time=60s --header="Cookie: $SANDBOXED" "http://localhost:3000/api/dashboard/10"` | Transactions: 136 hits<br>Availability: 100.00 %<br>Elapsed time: 60.07 secs<br>Response time: 6.30 secs<br>Transaction rate: 2.26 trans/sec<br>Throughput: 1.32 MB/sec<br>Longest transaction: 8.03<br>Shortest transaction: 4.85 |
| **Sandboxed - Without Tracking** | `/api/dashboard/10`                                | `siege --concurrent=15 --time=60s --header="Cookie: $SANDBOXED" "http://localhost:3000/api/dashboard/10"` | Transactions: 140 hits<br>Availability: 100.00 %<br>Elapsed time: 60.87 secs<br>Response time: 6.28 secs<br>Transaction rate: 2.30 trans/sec<br>Throughput: 1.34 MB/sec<br>Longest transaction: 8.06<br>Shortest transaction: 5.27 |
| **Sandboxed - With Tracking** | `/api/dashboard/10/query_metadata`                  | `siege --concurrent=15 --time=60s --header="Cookie: $SANDBOXED" "http://localhost:3000/api/dashboard/10/query_metadata"` | Transactions: 46 hits<br>Availability: 100.00 %<br>Elapsed time: 61.00 secs<br>Response time: 18.73 secs<br>Transaction rate: 0.75 trans/sec<br>Throughput: 0.40 MB/sec<br>Longest transaction: 20.51<br>Shortest transaction: 12.48 |
| **Sandboxed - Without Tracking** | `/api/dashboard/10/query_metadata`                  | `siege --concurrent=15 --time=60s --header="Cookie: $SANDBOXED" "http://localhost:3000/api/dashboard/10/query_metadata"` | Transactions: 37 hits<br>Availability: 100.00 %<br>Elapsed time: 60.96 secs<br>Response time: 19.98 secs<br>Transaction rate: 0.61 trans/sec<br>Throughput: 0.32 MB/sec<br>Longest transaction: 22.90<br>Shortest transaction: 12.86 |